fix: bottmo bar routing to profile page
